# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LocalAI Dev Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /workspaces/{workspaceId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileNode"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: Created file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileNode"
  /workspaces/{workspaceId}/files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file content
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWithContent"
    patch:
      operationId: updateFile
      tags: [files]
      summary: Update file content or metadata
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileNode"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── AI ────────────────────────────────────────────────
...
  /git/workspaces/{workspaceId}/status:
    get:
      operationId: getGitStatus
      tags: [git]
      summary: Get git status of a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Git status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitStatus"
  /git/workspaces/{workspaceId}/commits:
    get:
      operationId: getGitLog
      tags: [git]
      summary: Get recent git commits
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent commits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GitCommit"
    post:
      operationId: createCommit
      tags: [git]
      summary: Create a git commit
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitInput"
      responses:
        "200":
          description: Commit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitCommit"
  /git/workspaces/{workspaceId}/branches:
    get:
      operationId: listBranches
      tags: [git]
      summary: List git branches
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Branch list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchList"

  # ── Terminal Sessions ─────────────────────────────────
  /terminal/sessions:
    get:
      operationId: listTerminalSessions
      tags: [terminal]
      summary: List active terminal sessions
      responses:
        "200":
          description: Terminal sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TerminalSession"
    post:
      operationId: createTerminalSession
      tags: [terminal]
      summary: Create a terminal session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalSessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalSession"
  /terminal/sessions/{id}:
    delete:
      operationId: closeTerminalSession
      tags: [terminal]
      summary: Close a terminal session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Closed
components:
  schemas:
...
    Workspace:
      type: object
      required: [id, name, language, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        status:
          type: string
          enum: [idle, running, building, error]
        githubUrl:
          type: ["string", "null"]
        branch:
          type: ["string", "null"]
        containerId:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        lastActiveAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkspaceInput:
      type: object
      required: [name, language]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string
        githubUrl:
          type: string
        branch:
          type: string
    WorkspaceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        branch:
          type: string
...
    FileNode:
      type: object
      required: [id, workspaceId, name, path, type, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        language:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    FileWithContent:
      type: object
      required: [id, workspaceId, name, path, type, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        path:
          type: string
        type:
          type: string
        language:
          type: ["string", "null"]
        content:
          type: string
        size:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    FileInput:
      type: object
      required: [name, path, type]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
        language:
          type: string
    FileUpdate:
      type: object
      properties:
        content:
          type: string
        name:
          type: string
        path:
          type: string

    # ── AI Conversations ──────────────────────────────
...
    GitStatus:
      type: object
      required: [workspaceId, branch, staged, unstaged, untracked, ahead, behind]
      properties:
        workspaceId:
          type: integer
        branch:
          type: string
        staged:
          type: array
          items:
            type: string
        unstaged:
          type: array
          items:
            type: string
        untracked:
          type: array
          items:
            type: string
        ahead:
          type: integer
        behind:
          type: integer
    GitCommit:
      type: object
      required: [sha, message, author, date]
      properties:
        sha:
          type: string
        message:
          type: string
        author:
          type: string
        date:
          type: string
    CommitInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        files:
          type: array
          items:
            type: string
    BranchList:
      type: object
      required: [current, branches]
      properties:
        current:
          type: string
        branches:
          type: array
          items:
            type: string

    # ── Terminal ──────────────────────────────────────
    TerminalSession:
      type: object
      required: [id, workspaceId, name, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, closed]
        shell:
          type: string
        createdAt:
          type: string
    TerminalSessionInput:
      type: object
      required: [workspaceId, name]
      properties:
        workspaceId:
          type: integer
        name:
          type: string
        shell:
          type: string