# lib/api-zod/src/generated/types/fileWithContent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalAI Dev Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FileWithContent {
  id: number;
  workspaceId: number;
  name: string;
  path: string;
  type: string;
  /** @nullable */
  language?: string | null;
  content: string;
  /** @nullable */
  size?: number | null;
  createdAt: string;
  updatedAt: string;
}