# lib/api-zod/src/generated/types/gitStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalAI Dev Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GitStatus {
  workspaceId: number;
  branch: string;
  staged: string[];
  unstaged: string[];
  untracked: string[];
  ahead: number;
  behind: number;
}